workers/whisper-batch/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper-batch"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "openai-whisper>=20231117",
  "torch>=2.0",
]

[project.scripts]
whisper-batch = "whisper_batch.__main__:main"
whisper_batch = "whisper_batch.__main__:main"
